Adds changes for review comments changes
