+++ src/controller/RouteController.tsx
@@ -1,8 +1,8 @@
 
-import HeadMenu from "../components/headMenu";
+import Header from "../components/header";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import { BooksMain } from "../routes/books";
-import { CartMain } from "../routes/cart";
+import BooksMain from "../routes/books";
+import CartMain from "../routes/cart";
 
 function RouteController(): React.ReactElement {
