+++ src/routes/books/index.tsx
@@ -1,40 +1,3 @@
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { BookItem } from "../../components/bookItem";
-import HeadMenu from "../../components/headMenu";
-import { AppDispatch } from "../../store/store";
-import { Book, BooksState } from "../../types";
-import { booksSelector, getBooks } from "./book.reducer";
-import "./book.styles.css";
+import { BooksMain } from "./books.page";
 
-export const BooksMain = (): JSX.Element => {
-
-  const dispatch: AppDispatch = useDispatch();
-  const { books, loading } = useSelector(booksSelector);
-
-  useEffect(() => {
-    // fetch latest books
-    dispatch(getBooks());
-  }, []);
-
-  return (
-    <>
-      <HeadMenu />
-      <div className="books-page-wrapper">
-        {loading === "loading" ? (
-          <span className="book-loading">loading ...</span>
-        ) : (
-          <div className="books-page-container">
-            {books.map((book: Book) => (
-              <BookItem
-                {...book}
-                key={book.id}
-              />
-            ))}
-          </div>
-        )
-        }
-      </div>
-    </>
-  );
-};
+export default BooksMain;
