+++ src/routes/cart/index.tsx
@@ -1,40 +1,3 @@
-import { useSelector } from "react-redux";
-import { CartItemComponent } from "../../components/cartItem";
-import HeadMenu from "../../components/headMenu";
-import { CartItem } from "../../types";
-import { totalCartPrice } from "../../utils/utils";
-import { cartSelector } from "./cart.reducer";
-import "./cart.styles.css";
+import { CartMain } from "./cart.page";
 
-export const CartMain = (): JSX.Element => {
-
-  const { cart } = useSelector(cartSelector);
-
-  return (
-    <>
-      <HeadMenu />
-      <div className="cart-wrapper">
-        {!cart.length ? (
-          <span>uh oh! no books inside cart</span>
-        ) : (
-          <div className="cart-items-wrapper">
-            {cart.map((cartitem: CartItem) => (
-              <CartItemComponent
-                {...cartitem}
-                key={cartitem.id}
-              />
-            ))}
-            <div className="cart-total">
-              <div className="cart-total-div">
-                <strong className="cart-total-price">
-                  Total     ₹{totalCartPrice(cart)}
-                </strong>
-              </div>
-
-            </div>
-          </div>
-        )}
-      </div>
-    </>
-  );
-};
+export default CartMain;
